Python/MicrosoftFoundry.Classic/ModelContextProtocol/requirements.txt
agent-framework-core==1.0.0b251223
mcp[ws]>=1.23
azure-identity>=1